pyproject.toml
[project]
name = "kanomnom-python-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cachetools>=5.5.2",
    "diskcache>=5.6.3",
    "fastapi>=0.115.8",
    "flet[all]==0.28.3",
    "inflect>=7.5.0",
    "phonenumbers>=8.13.55",
    "polars>=1.34.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-extra-types>=2.10.2",
    "pydantic[email]>=2.10.6",
    "python-telegram-bot>=21.10",
    "rapidfuzz>=3.13.0",
    "requests>=2.32.4",
    "statsmodels>=0.14.5",
    "uvicorn[standard]>=0.34.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "matplotlib>=3.10.6",
    "optuna>=4.5.0",
    "pyright>=1.1.394",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.9.7",
]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 120
target-version = "py313"

[tool.pytest.ini_options]
markers = [
    "integration: marks tests that hit external services or take longer to run",
]
